Split range tests
